Move controllers to expected folder structure so Ember will automatically load them.
