+++ app/controllers/account/reset-success.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import AccountController from './account.js';
+import AccountController from '../account.js';
 
 export default AccountController.extend({
 });
